+++ exporters/otlp/otlpmetric/otlpmetrichttp/client_test.go
@@ -124,7 +124,7 @@ func TestConfig(t *testing.T) {
 		got := coll.Headers()
 		require.Regexp(t, "OTel Go OTLP over HTTP/protobuf metrics exporter/[01]\\..*", got)
 		require.Contains(t, got, key)
-		assert.Equal(t, got[key], []string{headers[key]})
+		assert.Equal(t, []string{headers[key]}, got[key])
 	})
 
 	t.Run("WithTimeout", func(t *testing.T) {
@@ -189,7 +189,7 @@ func TestConfig(t *testing.T) {
 		t.Cleanup(func() { close(rCh) })
 		t.Cleanup(func() { require.NoError(t, exp.Shutdown(ctx)) })
 		assert.NoError(t, exp.Export(ctx, &metricdata.ResourceMetrics{}), "failed retry")
-		assert.Len(t, rCh, 0, "failed HTTP responses did not occur")
+		assert.Empty(t, rCh, "failed HTTP responses did not occur")
 	})
 
 	t.Run("WithRetryAndExporterErr", func(t *testing.T) {
@@ -250,7 +250,7 @@ func TestConfig(t *testing.T) {
 
 		got := coll.Headers()
 		require.Contains(t, got, key)
-		assert.Equal(t, got[key], []string{headers[key]})
+		assert.Equal(t, []string{headers[key]}, got[key])
 	})
 
 	t.Run("WithProxy", func(t *testing.T) {
@@ -268,6 +268,6 @@ func TestConfig(t *testing.T) {
 
 		got := coll.Headers()
 		require.Contains(t, got, headerKeySetInProxy)
-		assert.Equal(t, got[headerKeySetInProxy], []string{headerValueSetInProxy})
+		assert.Equal(t, []string{headerValueSetInProxy}, got[headerKeySetInProxy])
 	})
 }