+++ trace/noop/noop_test.go
@@ -68,9 +68,9 @@ func assertAllExportedMethodNoPanic(rVal reflect.Value, rType reflect.Type) func
 
 func TestNewTracerProvider(t *testing.T) {
 	tp := NewTracerProvider()
-	assert.Equal(t, tp, TracerProvider{})
+	assert.Equal(t, TracerProvider{}, tp)
 	tracer := tp.Tracer("")
-	assert.Equal(t, tracer, Tracer{})
+	assert.Equal(t, Tracer{}, tracer)
 }
 
 func TestTracerStartPropagatesSpanContext(t *testing.T) {